fix: 시뮬레이션 확인 함수 분리
